# Hugo Migration (#8)
Switched from Svelte to Hugo. Svelte is great and I love it, but as
 this is a personal website/blog, I wanted to focus more on content
 creation rather than managing another website code, bugs, updates
 etc etc...

 ## removed github workflow

I realised I didn't really need to automate the deployment process since
the process is mostly automated by Hugo. I did create a local config
file with the deployment target, and added it to the gitignore file.

I'll probably look at automated deployments again in the future.